Setup `gaussian_filter_nan` in utils, use in `estimate_correlation_from_phase`
